python_server/requirements.txt
flask
pandas
konlpy
librosa
matplotlib
praat-parselmouth